refactor(react-ui-sfx, plugin-explorer): introduce FlockModel; factor out Visualization
react-ui-sfx
  Replace the `seeds` callback with a `FlockModel` class — an effect-atom-backed
  container for the boid array. Flock subscribes via `model.subscribe` and restarts
  the simulation whenever `setBoids` replaces the array; per-tick position changes
  are mutated in place on the same boid objects so external readers can sample the
  current positions via `model.boids` / `model.findBoid`.

  FlockBoid grows an optional `id` so callers seeding from a domain model (e.g.
  graph nodes) can match a boid back to its source on read.

  A `populate` callback replaces the old seeds shape — invoked once on first
  measurement when the model is empty.

plugin-explorer
  Factor `Visualization` out of `ExplorerArticle.tsx` into its own file; pull the
  variant union + toolbar metadata into `variants.ts`. ExplorerArticle is now a
  thin wrapper around the query editor, the variant toggle, and the DxAnchor
  preview dispatch.

  Visualization owns the FlockModel (one per mount, registry from
  RegistryContext). On entering swarm it seeds boids from `model.graph.nodes`,
  re-using positions from the last SVG layout where ids match. On leaving swarm
  it reads boid positions, translates back to center-origin, and writes them into
  `lastLayoutRef` so the next SVG projector animates from where the swarm ended.
