+++ packages/ui/react-ui-sfx/src/components/Flock/index.ts
@@ -3,3 +3,4 @@
 //
 
 export * from './Flock';
+export * from './FlockModel';
